Complex Capacity Expansion Notebook (07)
Publishing complex capacity expansion notebook, the final in our series.

Note that there is a known error in the provided data input: new build wind generators have incorrect capacity factor (1.0 in all periods). This will be patched shortly. For now, the results are odd but consistent with the input data, and the implementation works as intended.